Create GetParentSubformControl
Returns the subform control object that contains the subform that calls the procedure. This is useful if you use the same subform object in multiple subform controls. This code loops through all controls in the subform's parent form until it finds the subform control that contains the subform that called the procedure, and then returns that subform control object. The subform that calls the procedure simply passes the "Me" object to the procedure.